+++ src/components/LeftSidebar/GroupsList/GroupsList.vue
