+++ src/components/RightSidebar/Participants/CurrentParticipants/CurrentParticipants.vue
