+++ src/components/RightSidebar/Participants/ParticipantsList/Participant/Participant.vue
