+++ src/components/RightSidebar/Participants/ParticipantsList/ParticipantsList.vue
