+++ src/components/RightSidebar/Participants/ParticipantsTab.vue
@@ -54,7 +54,7 @@ import Caption from '../../Caption'
 import CurrentParticipants from './CurrentParticipants/CurrentParticipants'
 import Hint from '../../Hint'
 import ParticipantsList from './ParticipantsList/ParticipantsList'
-import SearchBox from '../../SearchBox'
+import SearchBox from '../../LeftSidebar/NewConversation/SearchBox'
 import debounce from 'debounce'
 import { EventBus } from '../../../services/EventBus'
 import { CONVERSATION, WEBINAR } from '../../../constants'
